// src/app/core/api/models/update-user-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Theme } from '../models/theme';
export interface UpdateUserRequest {
  email?: string;
  emailNotifications?: boolean;
  matchdayReminders?: boolean;
  preferredLanguage?: string;
  theme?: Theme;
  timezone?: string;
  twoFactorEnabled?: boolean;
}