// src/app/core/api/models/user.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { GlobalRole } from '../models/global-role';
import { Theme } from '../models/theme';
export interface User {
  active: boolean;
  createdAt: string;
  email?: string;
  emailNotifications?: boolean;
  emailVerified?: boolean;
  globalRole?: GlobalRole;
  id: number;
  matchdayReminders?: boolean;
  preferredLanguage?: string;
  theme?: Theme;
  timezone?: string;
  twoFactorEnabled?: boolean;
  username: string;
}